+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Releases
 
+## Version Release 1.1.6
+
+- Small README fix.
+
+## Version Release 1.1.5
+
+- Reworked README.
+- (Hopefully) fixed terminal credit de-sync between host and clients.
+- Legendary stamina upgrade still needs testing.
+
+## Version Release 1.1.0
+
+- Added legendary stamina upgrade.
+- Needs testing.
+
 ## Version Release 1.0.8
 
 - I hate not having upload previews.